receiver modules tests + moving of some folders

+++ WebService/WebService/Services/Data/Mongo/ReceiverModulesService.cs
@@ -5,7 +5,7 @@ using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using MongoDB.Driver;
-using WebService.Helpers.Extensions;
+using WebService.Helpers.Exceptions;
 using WebService.Models;
 
 namespace WebService.Services.Data.Mongo
@@ -15,21 +15,21 @@ namespace WebService.Services.Data.Mongo
     /// ReceiverModulesService is a class that extends from the <see cref="AMongoDataService{T}"/> class
     /// and by doing that implements the <see cref="IDataService{T}"/> interface.
     /// <para/>
-    /// It handles the saving and retreiving residents to and from the mongo database.
+    /// It handles the saving and retrieving residents to and from the mongo database.
     /// <para/>
-    /// The connectionstring, db name and collections that are used are stored in the IConfiguration dependency under the Database object.
+    /// The connection string, db name and collections that are used are stored in the IConfiguration dependency under the Database object.
     /// </summary>
-    public class ReceiverModulesService : AMongoDataService<ReceiverModule>, IReceiverModuleService
+    public class ReceiverModulesService : AMongoDataService<ReceiverModule>, IReceiverModulesService
     {
         /// <summary>
-        /// ReceiverModulesService is the contsructor to create an instance of the <see cref="ReceiverModulesService"/> class.
+        /// ReceiverModulesService is the constructor to create an instance of the <see cref="ReceiverModulesService"/> class.
         /// <para/>
-        /// The connectionstring, db name and collections that are used are stored in the IConfiguration dependency under the Database object.
+        /// The connection string, db name and collections that are used are stored in the IConfiguration dependency under the Database object.
         /// </summary>
         /// <param name="config"></param>
         public ReceiverModulesService(IConfiguration config)
         {
-            // create a new client and get the databas from it
+            // create a new client and get the database from it
             var db = new MongoClient(config["Database:ConnectionString"]).GetDatabase(config["Database:DatabaseName"]);
 
             // get the residents mongo collection
@@ -44,7 +44,7 @@ namespace WebService.Services.Data.Mongo
         public override IMongoCollection<ReceiverModule> MongoCollection { get; }
 
 
-        /// <inheritdoc cref="IReceiverModuleService.GetAsync(string,IEnumerable{Expression{Func{ReceiverModule,object}}})" />
+        /// <inheritdoc cref="IReceiverModulesService.GetAsync(string,IEnumerable{Expression{Func{ReceiverModule,object}}})" />
         /// <summary>
         /// GetAsync should return the receiver module with the given mac.
         /// </summary>
@@ -52,19 +52,25 @@ namespace WebService.Services.Data.Mongo
         /// <param name="propertiesToInclude">are the properties that should be included in the objects</param>
         /// <returns>The receiver module with the given mac</returns>
         public async Task<ReceiverModule> GetAsync(string mac,
-                IEnumerable<Expression<Func<ReceiverModule, object>>> propertiesToInclude = null)
+            IEnumerable<Expression<Func<ReceiverModule, object>>> propertiesToInclude = null)
         {
+            if (mac == null)
+                throw new ArgumentNullException(nameof(mac), "the mac address cannot be null");
+
             // get the item with the given id
-            var foundItem = MongoCollection.Find(x => x.Mac == mac);
+            var result = MongoCollection.Find(x => x.Mac == mac);
+
+            if (result.Count() <= 0)
+                throw new NotFoundException($"the {typeof(ReceiverModule).Name} with mac address {mac} was not found");
 
             // convert the properties to include to a list (if not null)
             var properties = propertiesToInclude?.ToList();
-            // if the proeprties are null or there are none, return all the properties
-            if (EnumerableExtensions.IsNullOrEmpty(properties))
-                return await foundItem.FirstOrDefaultAsync();
+            // if the properties are null or there are none, return all the properties
+            if (properties == null)
+                return await result.FirstOrDefaultAsync();
 
-            // create a propertyfilter
-            var selector = Builders<ReceiverModule>.Projection.Include(x => x.Id);
+            // create a property filter
+            var selector = Builders<ReceiverModule>.Projection.Include(x => x.Mac);
 
             //ReSharper disable once PossibleNullReferenceException
             // iterate over all the properties and add them to the filter
@@ -72,14 +78,14 @@ namespace WebService.Services.Data.Mongo
                 selector = selector.Include(property);
 
             // return the item
-            return await foundItem
+            return await result
                 // filter the properties
                 .Project<ReceiverModule>(selector)
                 // execute the query
                 .FirstOrDefaultAsync();
         }
 
-        /// <inheritdoc cref="IReceiverModuleService.RemoveAsync(string)" />
+        /// <inheritdoc cref="IReceiverModulesService.RemoveAsync(string)" />
         /// <summary>
         /// Remove removes the <see cref="ReceiverModule"/> with the given mac from the database.
         /// </summary>
@@ -90,10 +96,21 @@ namespace WebService.Services.Data.Mongo
         /// </returns>
         public async Task<bool> RemoveAsync(string mac)
         {
+            if (mac == null)
+                throw new ArgumentNullException(nameof(mac), "the mac address cannot be null");
+
             // remove the receiver module with the given mac from the database
             var result = await MongoCollection.DeleteOneAsync(x => x.Mac == mac);
-            // return true if something acutaly happened
-            return result.IsAcknowledged && result.DeletedCount > 0;
+
+            if (!result.IsAcknowledged)
+                throw new MongoException(
+                    $"the {typeof(ReceiverModule).Name} with mac address {mac} could not be removed from the db");
+
+            if (result.DeletedCount <= 0)
+                throw new NotFoundException($"the {typeof(ReceiverModule).Name} with mac address {mac} was not found");
+
+            // return true if something actually happened
+            return true;
         }
     }
 }
